pyproject.toml
[tool.poetry]
name = "todo-project"
version = "0.1.0"
description = "Простой CLI todo менеджер"
authors = ["Maks <maks267361@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.13"
sqlalchemy = "*"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
tqdm = "^4.67.1"
alembic = "^1.16.5"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.3"
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
coverage = "^7.9.2"
pytest-asyncio = "^1.1.0"
httpx = ">=0.25"
pytest-dotenv = "^0.5.2"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
fix = true

[tool.ruff.lint]
extend-select = ["E", "W", "F"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
plugins = [
  "sqlalchemy.ext.mypy.plugin",
  "pydantic.mypy"
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term --cov-report=html"
